+++ spec/edge_cases_integration_spec.rb
@@ -5,7 +5,6 @@ require "spec_helper"
 class EdgeProduct < ActiveRecord::Base
   include ParadeDB::Model
   self.table_name = :products
-  self.has_paradedb_index = true
 end
 
 RSpec.describe "EdgeCasesIntegrationTest" do